app/requirements.txt
click==8.3.0
colorama==0.4.6
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyodbc==5.2.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
alembic==1.16.5
blinker==1.9.0
Mako==1.3.10
python-dotenv==1.1.1
bcrypt==5.0.0
Flask-JWT-Extended==4.7.1
PyJWT==2.10.1
tzdata==2025.2
reportlab==4.4.4